chore: 사용하지 않는 파일 삭제 및 리소스 정리 - 이전에 구현하고 현재 사용하지 않는 스크린 삭제
