+++ features/carefullcontents/src/main/java/com/cases/carefull/features/carefullcontents/routine/DietScreen.kt
@@ -198,142 +198,3 @@ fun FoodItemRow(food: MealRecord, onRemove: () -> Unit) {
 		}
 	}
 }
-
-
-//@Composable
-//fun DietScreen(viewModel: DietViewModel) {
-//	val uiState = viewModel.uiState.collectAsState()
-//
-//	Column {
-//		Spacer(modifier = Modifier.height(24.dp))
-//		Text(
-//			text = "오늘 식사 내역", style = MaterialTheme.typography.bodyLarge,
-//			modifier = Modifier
-//				.fillMaxWidth()
-//				.padding(horizontal = 16.dp),
-//			textAlign = TextAlign.Center
-//		)
-//		Text(
-//			text = "총 ${uiState.value.totalCalories} 칼로리",
-//			style = MaterialTheme.typography.bodyLarge,
-//			modifier = Modifier
-//				.fillMaxWidth()
-//				.padding(horizontal = 16.dp),
-//			textAlign = TextAlign.Center
-//		)
-//		LazyColumn(
-//			modifier = Modifier
-//				.fillMaxSize()
-//				.padding(16.dp)
-//		) {
-//			item {
-//				MealType.entries.forEach { mealType ->
-//					MealSection(
-//						mealType = mealType,
-//						addedFoods = uiState.value.mealsByTime[mealType] ?: emptyList(),
-//						onCameraClick = {},
-//						onAddClick = {
-//						},
-//						onRemoveClick = {}
-//
-//					)
-//				}
-//			}
-//		}
-//	}
-//}
-//
-//
-//
-//@Composable
-//fun MealSection(
-//	mealType: MealType,
-//	addedFoods: List<MealRecord>,
-//	onCameraClick: () -> Unit,
-//	onAddClick: () -> Unit,
-//	onRemoveClick: () -> Unit
-//) {
-//	Card(
-//		modifier = Modifier
-//			.fillMaxWidth()
-//			.padding(vertical = 12.dp),
-//		elevation = CardDefaults.cardElevation(defaultElevation = 2.dp)
-//	) {
-//		Column(
-//			modifier = Modifier
-//				.fillMaxWidth()
-//		) {
-//			Row(
-//				modifier = Modifier
-//					.fillMaxWidth()
-//					.padding(8.dp),
-//				verticalAlignment = Alignment.CenterVertically
-//			) {
-//				Text(
-//					text = mealType.time,
-//					style = MaterialTheme.typography.bodyLarge,
-//					fontWeight = FontWeight.Bold
-//				)
-//				Spacer(modifier = Modifier.weight(1f))
-//				IconButton(onClick = onCameraClick) {
-//					Icon(
-//						imageVector = Icons.Default.Camera,
-//						contentDescription = "${mealType.time} 음식 추가",
-//						tint = MaterialTheme.colorScheme.primary,
-//						modifier = Modifier.size(28.dp)
-//					)
-//				}
-//				IconButton(onClick = onAddClick) {
-//					Icon(
-//						imageVector = Icons.Default.Add,
-//						contentDescription = "${mealType.time} 음식 검색",
-//						tint = MaterialTheme.colorScheme.primary,
-//						modifier = Modifier.size(28.dp)
-//					)
-//				}
-//			}
-//			HorizontalDivider(
-//				thickness = 1.dp,
-//				color = Color.Black
-//			)
-//			if (addedFoods.isNotEmpty()) {
-//				Column(
-//					modifier = Modifier.fillMaxWidth(),
-//					verticalArrangement = Arrangement.spacedBy(8.dp)
-//				) {
-//					addedFoods.forEach { addedFood ->
-//						FoodItemRow(
-//							food = addedFood,
-//							onRemove = onRemoveClick
-//						)
-//					}
-//				}
-//			} else {
-//				Text(
-//					text = "아직 추가된 음식이 없습니다.",
-//					style = MaterialTheme.typography.bodyMedium,
-//					color = Color.Gray,
-//					modifier = Modifier.padding(8.dp)
-//				)
-//			}
-//		}
-//	}
-//}
-//
-//@Composable
-//fun FoodItemRow(food: MealRecord, onRemove: () -> Unit) {
-//	Row(
-//		modifier = Modifier
-//			.fillMaxWidth()
-//			.padding(horizontal = 16.dp, vertical = 8.dp),
-//		verticalAlignment = Alignment.CenterVertically
-//	) {
-//		Text(
-//			text = "${food.name} (${food.calories} kcal)",
-//			modifier = Modifier.weight(1f)
-//		)
-//		IconButton(onClick = onRemove, modifier = Modifier.size(20.dp)) {
-//			Icon(imageVector = Icons.Default.Close, contentDescription = "삭제")
-//		}
-//	}
-//}
